Fixed an issue where re-executing LOAD DATASET might not re-load the file
LOAD DATASET goes through several checks to avoid re-loading a dataset when it doesn't have to.
Specifically, if Vistrails metadata already exists for the file in question, it then checks to see
if Mimir has already loaded the file as a dataset.  If Mimir can't access the dataset, Vizier
prompts Mimir to reload the file.

Previously, Vistrails would skip collecting some of the metadata that Mimir needed to reload the
file (i.e., the file URL) if Vistrails metadata already existed.  Now, the metadata is also
collected if the dataset is absent from Mimir.