Adding support for a workflow sql endpoint for use in support tools
This PR adds support for a new endpoint
  /projects/<string:project_id>/branches/<string:branch_id>/head/sql
The endpoint takes a single parameter that may be passed unescaped in the request body or by the `query` GET argument.  Both GET and POST methods are valid.

With the Mimir backend, this endpoint now allows running SQL queries against the datasets at the end of a workflow.  The primary immediate use is to allow Vizier to act as the datasource for a data visualization tool that we're building.

DEPENDS ON: https://github.com/UBOdin/mimir-api/pull/23