Adding support for profile-on-demand
A bunch of files got modified because Vizier is a giant soup of redirections, but the fundamental change is simple.  the /projects/<projectid>/datasets/<datasetid> route now has an optional `profile` GET argument.  When invoked with ?profile=true, the result is guaranteed to include profiler-generated properties.  When not invoked with ?profile=true, profiler properties may or may not be present (they're cached, but not re-generated if the request does not come in).  The presence of profiling properties can be tested for by checking for the "is_profiled" property (i.e., $.properties.is_profiled)